DockerDemo/requirements.txt
langchain==0.0.225
streamlit==1.24.0
streamlit-chat==0.1.1
sentence-transformers==2.2.2

